requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.4.1; python_version >= '3.6'
django-js-asset==1.2.2
django-mptt==0.13.3
django==3.2.7
pillow==8.3.2
psycopg2==2.9.1
pytz==2021.1
sqlparse==0.4.2; python_version >= '3.5'
